This commit adds an "Add64" method on the progress bar
This is for consistancy's sake.  There is a "New64" method that allows
the library consumer to create an int64-based progress bar, but
currently it only allows type int to use the "Add" method to increase
the increment.  This is strange because there is casting of a regular
int to an int64 behind the scenes to accomodate this.

This commit adds an Add64 method directly, to minimize casting.  That
way all values sent to the ProgressBar are of type int64.