use utf8.RuneCountInString() instead of builtin len()
When we have non-english characters or utf-8 symbols in Prefix or Postfix the effective width
calculation turns out to be wrong. This is a limitation of built-in len() function which returns
the number of bytes instead of the number of characters to be printed on console.

Number of characters is perhaps what is needed for proper barWidth calculation.  So it is better
to use utf8.RuneCountInString() to calculate the best possible column length used by a given utf-8 character.