Add windows code, ts.GetSize() returned wrong value for columns.
in ``ts.GetSize()`` - bufferScreenInfo.srWindow.Right is not really
the column width but instead its.

```The x-coordinate of the lower right corner of the rectangle.```

Also to have an actual value of the column one needs to do

`srWindow.Right - srWindow.Left + 1`

Rather than that use `dwSize` instead which in-fact contains the screen
buffer size in character columns and rows.