Improve the portability of termios handling.
Previously this was using the `syscall` package.  All of the helper
functions that were likely missing at the time this was written are
now available under the `golang.org/x/sys/unix` package.

Hat-tip to `golang.org/x/crypto/ssh/terminal` for having most of
the necessary bits already stubbed out and available for use.