Add 'reset' for more flexibility
There are cases when the state transition from the bar being "started" to "finished" cause an over-application of the last set color in the bar.

For instance: 

```
`{{ blue "no reset" }} {{ counters . | red }} {{bar . "|" (green "█" ) ">" "-" "|"}} {{percent . }} {{speed .}} {{rtime .}}`
``` 

causes everything after `green` to be green. Once we add `reset` there's a slight improvement during the render, as the trailing elements are not green, but upon transition to the "finished" state, the rest of the line turns green:

```
`{{ blue "nested reset" }} {{ counters . | red }} {{bar . "|" (green "█" ) (reset ">") "-" "|"}} {{percent . }} {{speed .}} {{rtime .}}`
 ```

and finally, with an external reset:

```
`{{ blue "external reset" }} {{ counters . | red }} {{bar . "|" (green "█" ) ">" "-" "|"}} {{percent . | reset}} {{speed .}} {{rtime .}}`
```
which functions very similar to the way that terminal color resets work in shells.

There may be a better way to manage the state transition form "started" to "finished", but this does the trick.