Reduce resolution for elapsed time > 1s
db60ccefe5cb616c2f8a1bde35ee059118dce62a increased the resolution shown
on elapsed time from seconds to milliseconds. This applied regardless of
the total duration, so even if the duration was large, a high resolution
would be displayed, such as "897.213s". This patch changes the behavior
so that milliseconds are only shown if the total time is less than a
second, so that example would be displayed as "897s" instead.