Add parameter to specify elapsed time precision
db60ccefe5cb616c2f8a1bde35ee059118dce62a (released in v3.0.7) increased
the resolution shown on elapsed time from seconds to milliseconds. The
behavior prior to v3.0.7 was to truncate the time to the nearest second.
This patch allows the user to control the precision. The default is to
round to the nearest second, like the pre-v3.0.7 behavior. This patch
also switches from truncation to rounding.